requirements.txt
Twisted==15.1.0
zope.interface==4.1.2
PyYAML==3.11